src/rc_car/requirements.txt
paho-mqtt==2.*
flask==3.*
picamera2 
